+++ Thunderstore-EN/CHANGELOG.md
@@ -1,6 +1,17 @@
-Super super small release, should've included it in 1.0.5 to be honest.
+I have never felt this guilty before. As compensation, there's nothing ;P
 
 Changes:
-- Changed the sprite of 'Volcanic Shard'
+- 'Tainted Red Scarf' now properly evolves (double triple quadruple checked)
+- 'Shrinking Potion' now properly evolves
+- 'Frozen Spear' can now apply freeze with normal attacks as well
+- Freeze chance of 'Spear of the Frozen Moon' is increased from 15% to 20%, and is now able to apply freeze with normal attacks as well
+- Increased wound chance upon crit of 'Blood-Soaked Javelin' from 5% to 15%
+- Decreased crit chance while midair of 'Rotten Wings' from 15% to 12%
+- 'Tainted Finger' and 'Recovered Fingers' both now amplify skill damage dealt to enemies by 25% and increases magic attack by 60%
+- 'Tainted Finger' now increases incoming damage by 40%
+- 'Recovered Fingers' now decreases incoming damage by 10%
+- 'Corrupted God's Hand' now amplifies skill damage dealt to enemies by 55%, increases magic attack by 90%, and increases incoming damage by 40%
+- Fixed the Korean description of 'Recovered Fingers'
+- 'Recovered Fingers' no longer consume 'Grace of Leonia' when evolved
 
-I hope this is the last change I need. Please, I believe it's time I move on.
+It's time to pray to the holy devs to not make me do another update because of the new update.